Split `GitFileManager` too
