+++ src/Microsoft.DotNet.Darc/DarcLib/Actions/Local.cs
@@ -15,7 +15,7 @@ namespace Microsoft.DotNet.DarcLib;
 
 public class Local : ILocal
 {
-    private readonly GitFileManager _fileManager;
+    private readonly LibGit2FileManager _fileManager;
     private readonly ILocalLibGit2Client _gitClient;
     private readonly IVersionDetailsParser _versionDetailsParser;
 
@@ -34,7 +34,7 @@ public class Local : ILocal
         _logger = logger;
         _versionDetailsParser = new VersionDetailsParser();
         _gitClient = new LocalLibGit2Client(remoteConfiguration, new ProcessManager(logger, GitExecutable), logger);
-        _fileManager = new GitFileManager(_gitClient, _versionDetailsParser, logger);
+        _fileManager = new LibGit2FileManager(_gitClient, _versionDetailsParser, logger);
 
         _repoRootDir = new(() => overrideRootPath ?? _gitClient.GetRootDirAsync().GetAwaiter().GetResult(), LazyThreadSafetyMode.PublicationOnly);
     }
