--- src/Microsoft.DotNet.Darc/DarcLib/DependencyOperations.cs
@@ -1,80 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Threading.Tasks;
-
-namespace Microsoft.DotNet.DarcLib;
-
-public static class DependencyOperations
-{
-    public static readonly Dictionary<string, KnownDependencyType> _knownAssetNames =
-        new Dictionary<string, KnownDependencyType>
-        {
-            {"Microsoft.DotNet.Arcade.Sdk", KnownDependencyType.GlobalJson},
-            {"Microsoft.DotNet.Build.Tasks.SharedFramework.Sdk", KnownDependencyType.GlobalJson},
-            {"Microsoft.DotNet.Helix.Sdk", KnownDependencyType.GlobalJson},
-            {"Microsoft.DotNet.SharedFramework.Sdk", KnownDependencyType.GlobalJson},
-            {"Microsoft.NET.SharedFramework.Sdk", KnownDependencyType.GlobalJson},
-            {"Microsoft.DotNet.CMake.Sdk", KnownDependencyType.GlobalJson},
-            {"dotnet", KnownDependencyType.GlobalJson}
-        };
-
-    private static readonly Dictionary<KnownDependencyType, Delegate> _dependenciesToFuncMapping =
-        new Dictionary<KnownDependencyType, Delegate>
-        {
-            {
-                KnownDependencyType.GlobalJson,
-                new Func<GitFileManager, string, string, DependencyDetail, Task>(UpdateGlobalJson)
-            }
-        };
-
-    public static bool TryGetKnownUpdater(string name, out Delegate function)
-    {
-        if (_knownAssetNames.ContainsKey(name))
-        {
-            function = _dependenciesToFuncMapping[_knownAssetNames[name]];
-            return true;
-        }
-
-        function = null;
-        return false;
-    }
-
-    public static async Task UpdateGlobalJson(
-        GitFileManager fileManager,
-        string repository,
-        string branch,
-        DependencyDetail dependency)
-    {
-        var dependencyMapping = new Dictionary<string, string>
-        {
-            {"Microsoft.DotNet.Arcade.Sdk", "msbuild-sdks"},
-            {"Microsoft.DotNet.Build.Tasks.SharedFramework.Sdk", "msbuild-sdks"},
-            {"Microsoft.DotNet.Helix.Sdk", "msbuild-sdks"},
-            {"Microsoft.DotNet.SharedFramework.Sdk", "msbuild-sdks"},
-            {"Microsoft.NET.SharedFramework.Sdk", "msbuild-sdks"},
-            {"dotnet", "tools"}
-        };
-
-        if (!dependencyMapping.ContainsKey(dependency.Name))
-        {
-            throw new Exception($"Dependency '{dependency.Name}' has no parent mapping defined.");
-        }
-
-        string parent = dependencyMapping[dependency.Name];
-
-        await fileManager.AddDependencyToGlobalJson(
-            repository,
-            branch,
-            parent,
-            dependency.Name,
-            dependency.Version);
-        await fileManager.AddDependencyToVersionDetailsAsync(
-            repository,
-            branch,
-            dependency);
-    }
-}
